requirements.txt
# الأساسيات الرياضية والمعالجة
numpy>=1.24.0
pandas>=2.0.0

# استخراج البيانات من PDF
pymupdf>=1.23.0
pdfplumber>=0.10.0

# المحرك الدلالي والذكاء الاصطناعي
sentence-transformers>=2.2.2
torch>=2.0.0
python-dotenv>=1.0.0

# أدوات النظام
tqdm>=4.65.0